+++ starkbank/dynamicbrcode/rule/__rule.py
@@ -1,0 +1,11 @@
+from starkcore.utils.subresource import SubResource
+
+
+class Rule(SubResource):
+
+    def __init__(self, key, value):
+        self.key = key
+        self.value = value
+
+
+_sub_resource = {"class": Rule, "name": "Rule"}